Replace type italics with inline code blocks
In the case of actual types used in the language. I noticed some of the
types where wrapped in ` and some in *. I replaced all the types wrapped
in * with `. I think this makes it easier to read which names are actual
types used in the language and which are not.